Place picture frames directly on the floor and disable shadows. (#14)
The picture frames were floating in mid-air, and fall to the ground during the first few seconds due to gravity in the simulation. This change places them directly on the ground.

Also, having shadows enabled caused the scene to be dark when there are no directional lights. Since we don't have directional lights in the scene, I disabled the shadows so that the scene will be rendered with the correct brightness.

Tested manually loading the scene in gzclient and observe picture frames are on the floor, and the scene is bright.